# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RevenueDog API",
    "version": "0.1.0",
    "description": "Open-source, self-hostable subscription & in-app-purchase management API (a RevenueCat alternative). Public endpoints (`/v1/*`) use a per-app public key (`pk_`); admin endpoints (`/v1/admin/*`) use a secret key (`sk_`).",
    "license": { "name": "MIT", "url": "https://opensource.org/licenses/MIT" }
  },
  "servers": [{ "url": "http://localhost:8787", "description": "Local development" }],
  "security": [{ "secretKey": [] }],
...
  "paths": {
...
    "/v1/admin/products": {
      "get": { "tags": ["Admin: Products"], "summary": "List products", "responses": { "200": { "$ref": "#/components/responses/ProductList" } } },
      "post": { "tags": ["Admin: Products"], "summary": "Create product", "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProductCreate" } } } }, "responses": { "201": { "$ref": "#/components/responses/Product" }, "409": { "$ref": "#/components/responses/Error" } } }
    },
    "/v1/admin/products/import": {
      "post": { "tags": ["Admin: Products"], "summary": "Bulk import products", "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["products"], "properties": { "products": { "type": "array", "items": { "$ref": "#/components/schemas/ProductCreate" } } } } } } }, "responses": { "200": { "description": "Import result", "content": { "application/json": { "schema": { "type": "object", "properties": { "imported": { "type": "integer" }, "skipped": { "type": "integer" }, "failed": { "type": "array", "items": { "type": "object" } } } } } } } } }
    },
    "/v1/admin/products/import/{store}": {
      "post": { "tags": ["Admin: Products"], "summary": "Import catalog from store API", "parameters": [{ "name": "store", "in": "path", "required": true, "schema": { "type": "string", "enum": ["app_store", "play_store"] } }], "responses": { "200": { "description": "Import result" }, "502": { "$ref": "#/components/responses/Error" } } }
    },
    "/v1/admin/products/{id}": {
      "parameters": [{ "$ref": "#/components/parameters/Id" }],
      "get": { "tags": ["Admin: Products"], "summary": "Get product", "responses": { "200": { "$ref": "#/components/responses/Product" }, "404": { "$ref": "#/components/responses/Error" } } },
      "patch": { "tags": ["Admin: Products"], "summary": "Update product", "requestBody": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProductCreate" } } } }, "responses": { "200": { "$ref": "#/components/responses/Product" } } },
      "delete": { "tags": ["Admin: Products"], "summary": "Delete product", "responses": { "200": { "$ref": "#/components/responses/Ok" }, "404": { "$ref": "#/components/responses/Error" } } }
    },
    "/v1/admin/entitlements": {
      "get": { "tags": ["Admin: Entitlements"], "summary": "List entitlements", "responses": { "200": { "description": "List", "content": { "application/json": { "schema": { "type": "object", "properties": { "items": { "type": "array", "items": { "$ref": "#/components/schemas/Entitlement" } } } } } } } } },
      "post": { "tags": ["Admin: Entitlements"], "summary": "Create entitlement", "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/EntitlementCreate" } } } }, "responses": { "201": { "description": "Created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Entitlement" } } } } } }
    },
    "/v1/admin/entitlements/{id}": {
      "parameters": [{ "$ref": "#/components/parameters/Id" }],
      "get": { "tags": ["Admin: Entitlements"], "summary": "Get entitlement", "responses": { "200": { "description": "Entitlement", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Entitlement" } } } }, "404": { "$ref": "#/components/responses/Error" } } },
      "patch": { "tags": ["Admin: Entitlements"], "summary": "Update entitlement", "requestBody": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/EntitlementCreate" } } } }, "responses": { "200": { "description": "Updated", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Entitlement" } } } } } },
      "delete": { "tags": ["Admin: Entitlements"], "summary": "Delete entitlement", "responses": { "200": { "$ref": "#/components/responses/Ok" } } }
    },
    "/v1/admin/offerings": {
      "get": { "tags": ["Admin: Offerings"], "summary": "List offerings", "responses": { "200": { "description": "List", "content": { "application/json": { "schema": { "type": "object", "properties": { "items": { "type": "array", "items": { "$ref": "#/components/schemas/Offering" } } } } } } } } },
      "post": { "tags": ["Admin: Offerings"], "summary": "Create offering", "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/OfferingCreate" } } } }, "responses": { "201": { "description": "Created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Offering" } } } } } }
    },
    "/v1/admin/offerings/{id}": {
      "parameters": [{ "$ref": "#/components/parameters/Id" }],
      "get": { "tags": ["Admin: Offerings"], "summary": "Get offering", "responses": { "200": { "description": "Offering", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Offering" } } } } } },
      "patch": { "tags": ["Admin: Offerings"], "summary": "Update offering", "requestBody": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/OfferingCreate" } } } }, "responses": { "200": { "description": "Updated", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Offering" } } } } } },
      "delete": { "tags": ["Admin: Offerings"], "summary": "Delete offering", "responses": { "200": { "$ref": "#/components/responses/Ok" } } }
    },
    "/v1/admin/experiments": {
      "get": { "tags": ["Admin: Experiments"], "summary": "List experiments", "responses": { "200": { "description": "List", "content": { "application/json": { "schema": { "type": "object", "properties": { "items": { "type": "array", "items": { "$ref": "#/components/schemas/Experiment" } } } } } } } } },
      "post": { "tags": ["Admin: Experiments"], "summary": "Create experiment", "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ExperimentCreate" } } } }, "responses": { "201": { "description": "Created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Experiment" } } } } } }
    },
    "/v1/admin/experiments/{id}": {
      "parameters": [{ "$ref": "#/components/parameters/Id" }],
      "get": { "tags": ["Admin: Experiments"], "summary": "Get experiment", "responses": { "200": { "description": "Experiment", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Experiment" } } } } } },
      "patch": { "tags": ["Admin: Experiments"], "summary": "Update experiment", "requestBody": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ExperimentCreate" } } } }, "responses": { "200": { "description": "Updated", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Experiment" } } } } } },
      "delete": { "tags": ["Admin: Experiments"], "summary": "Delete experiment", "responses": { "200": { "$ref": "#/components/responses/Ok" } } }
    },
    "/v1/admin/experiments/{id}/stop": {
      "parameters": [{ "$ref": "#/components/parameters/Id" }],
      "post": { "tags": ["Admin: Experiments"], "summary": "Stop experiment", "responses": { "200": { "description": "Stopped", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Experiment" } } } } } }
    },
    "/v1/admin/experiments/{id}/results": {
      "parameters": [{ "$ref": "#/components/parameters/Id" }],
      "get": { "tags": ["Admin: Experiments"], "summary": "Experiment results", "responses": { "200": { "description": "Per-variant enrollment/conversion", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ExperimentResults" } } } } } }
    },
    "/v1/admin/apps": {
      "get": { "tags": ["Admin: Apps"], "summary": "List apps", "responses": { "200": { "description": "List", "content": { "application/json": { "schema": { "type": "object", "properties": { "items": { "type": "array", "items": { "$ref": "#/components/schemas/App" } } } } } } } } },
      "post": { "tags": ["Admin: Apps"], "summary": "Create app (generates a pk_ key)", "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["name"], "properties": { "name": { "type": "string" }, "bundle_id": { "type": ["string", "null"] }, "package_name": { "type": ["string", "null"] } } } } } }, "responses": { "201": { "description": "Created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/App" } } } } } }
    },
    "/v1/admin/apps/{id}": {
      "parameters": [{ "$ref": "#/components/parameters/Id" }],
      "get": { "tags": ["Admin: Apps"], "summary": "Get app", "responses": { "200": { "description": "App", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/App" } } } } } },
      "delete": { "tags": ["Admin: Apps"], "summary": "Delete app", "responses": { "200": { "$ref": "#/components/responses/Ok" } } }
    },
    "/v1/admin/imports": {
      "post": { "tags": ["Admin: Subscribers"], "summary": "Bulk import historical receipts", "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["receipts"], "properties": { "receipts": { "type": "array", "items": { "$ref": "#/components/schemas/ReceiptRequest" } } } } } } }, "responses": { "200": { "description": "Import result", "content": { "application/json": { "schema": { "type": "object", "properties": { "imported": { "type": "integer" }, "failed": { "type": "array", "items": { "type": "object" } } } } } } } } }
    },
    "/v1/admin/subscribers": {
      "get": { "tags": ["Admin: Subscribers"], "summary": "List subscribers", "parameters": [{ "name": "limit", "in": "query", "schema": { "type": "integer" } }, { "name": "offset", "in": "query", "schema": { "type": "integer" } }], "responses": { "200": { "description": "List", "content": { "application/json": { "schema": { "type": "object", "properties": { "items": { "type": "array", "items": { "$ref": "#/components/schemas/SubscriberSummary" } } } } } } } } }
    },
    "/v1/admin/subscribers/{app_user_id}": {
      "parameters": [{ "$ref": "#/components/parameters/AppUserId" }],
      "get": { "tags": ["Admin: Subscribers"], "summary": "Get subscriber CustomerInfo", "responses": { "200": { "$ref": "#/components/responses/CustomerInfo" }, "404": { "$ref": "#/components/responses/Error" } } }
    },
    "/v1/admin/subscribers/{app_user_id}/entitlements/{identifier}/grant": {
      "parameters": [{ "$ref": "#/components/parameters/AppUserId" }, { "name": "identifier", "in": "path", "required": true, "schema": { "type": "string" } }],
      "post": { "tags": ["Admin: Subscribers"], "summary": "Grant a promotional entitlement", "requestBody": { "content": { "application/json": { "schema": { "type": "object", "properties": { "expires_date": { "type": ["string", "null"], "format": "date-time" } } } } } }, "responses": { "200": { "$ref": "#/components/responses/CustomerInfo" } } }
    },
    "/v1/admin/subscribers/{app_user_id}/entitlements/{identifier}/revoke": {
      "parameters": [{ "$ref": "#/components/parameters/AppUserId" }, { "name": "identifier", "in": "path", "required": true, "schema": { "type": "string" } }],
      "post": { "tags": ["Admin: Subscribers"], "summary": "Revoke a promotional entitlement", "responses": { "200": { "$ref": "#/components/responses/CustomerInfo" } } }
    },
    "/v1/admin/webhooks": {
      "get": { "tags": ["Admin: Webhooks"], "summary": "List webhooks", "responses": { "200": { "description": "List", "content": { "application/json": { "schema": { "type": "object", "properties": { "items": { "type": "array", "items": { "$ref": "#/components/schemas/Webhook" } } } } } } } } },
      "post": { "tags": ["Admin: Webhooks"], "summary": "Create webhook", "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["url"], "properties": { "url": { "type": "string", "format": "uri" }, "events": { "oneOf": [{ "type": "string", "enum": ["*"] }, { "type": "array", "items": { "type": "string" } }] }, "active": { "type": "boolean" } } } } } }, "responses": { "201": { "description": "Created (includes signing secret)", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Webhook" } } } } } }
    },
    "/v1/admin/webhooks/{id}": {
      "parameters": [{ "$ref": "#/components/parameters/Id" }],
      "get": { "tags": ["Admin: Webhooks"], "summary": "Get webhook", "responses": { "200": { "description": "Webhook", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Webhook" } } } } } },
      "patch": { "tags": ["Admin: Webhooks"], "summary": "Update webhook", "requestBody": { "content": { "application/json": { "schema": { "type": "object" } } } }, "responses": { "200": { "description": "Updated", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Webhook" } } } } } },
      "delete": { "tags": ["Admin: Webhooks"], "summary": "Delete webhook", "responses": { "200": { "$ref": "#/components/responses/Ok" } } }
    },
...
    "/v1/admin/webhooks/{id}/test": {
      "parameters": [{ "$ref": "#/components/parameters/Id" }],
      "post": { "tags": ["Admin: Webhooks"], "summary": "Dispatch a test event", "responses": { "200": { "$ref": "#/components/responses/Ok" } } }
    },
    "/v1/admin/events": {
      "get": { "tags": ["Admin: Analytics"], "summary": "Recent events feed", "parameters": [{ "name": "limit", "in": "query", "schema": { "type": "integer" } }], "responses": { "200": { "description": "Events", "content": { "application/json": { "schema": { "type": "object", "properties": { "items": { "type": "array", "items": { "type": "object" } } } } } } } } }
    },
...
    "/v1/admin/overview": {
      "get": { "tags": ["Admin: Analytics"], "summary": "Dashboard overview metrics", "parameters": [{ "name": "range", "in": "query", "schema": { "type": "integer", "default": 28 } }], "responses": { "200": { "description": "Overview", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Overview" } } } } } }
    },
...
  },
  "components": {
    "securitySchemes": {
      "publicKey": { "type": "http", "scheme": "bearer", "description": "App public key (pk_). Safe to ship in mobile apps." },
      "secretKey": { "type": "http", "scheme": "bearer", "description": "Secret key (sk_). Server/dashboard only." }
    },
    "parameters": {
      "AppUserId": { "name": "app_user_id", "in": "path", "required": true, "schema": { "type": "string" } },
      "Id": { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } },
...
    },
    "responses": {
      "Ok": { "description": "Success", "content": { "application/json": { "schema": { "type": "object", "properties": { "ok": { "type": "boolean" } } } } } },
      "Error": { "description": "Error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
      "CustomerInfo": { "description": "CustomerInfo", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CustomerInfo" } } } },
      "Product": { "description": "Product", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Product" } } } },
      "ProductList": { "description": "Product list", "content": { "application/json": { "schema": { "type": "object", "properties": { "items": { "type": "array", "items": { "$ref": "#/components/schemas/Product" } } } } } } }
    },
    "schemas": {
...
      "ReceiptRequest": {
        "type": "object",
        "required": ["app_user_id", "fetch_token", "product_id", "store"],
        "properties": {
          "app_user_id": { "type": "string" },
          "fetch_token": { "type": "string" },
          "product_id": { "type": "string" },
          "store": { "type": "string", "enum": ["app_store", "play_store"] },
          "presented_offering_identifier": { "type": ["string", "null"] },
          "price": { "type": ["number", "null"] },
          "currency": { "type": ["string", "null"] }
        }
      },
...
      "ProductCreate": {
        "type": "object",
        "required": ["store_identifier", "type", "store", "display_name"],
        "properties": {
          "store_identifier": { "type": "string" },
          "type": { "type": "string", "enum": ["subscription", "non_consumable", "consumable"] },
          "store": { "type": "string", "enum": ["app_store", "play_store"] },
          "display_name": { "type": "string" },
          "duration": { "type": ["string", "null"], "description": "ISO-8601 duration, e.g. P1M" }
        }
      },
...
      "EntitlementCreate": {
        "type": "object",
        "required": ["identifier", "display_name"],
        "properties": { "identifier": { "type": "string" }, "display_name": { "type": "string" }, "product_ids": { "type": "array", "items": { "type": "string" } } }
      },
      "Entitlement": { "allOf": [{ "type": "object", "properties": { "id": { "type": "string" } } }, { "$ref": "#/components/schemas/EntitlementCreate" }] },
...
      "OfferingCreate": {
        "type": "object",
        "required": ["identifier"],
        "properties": {
          "identifier": { "type": "string" },
          "description": { "type": "string" },
          "metadata": { "type": "object" },
          "is_current": { "type": "boolean" },
          "packages": { "type": "array", "items": { "$ref": "#/components/schemas/OfferingPackage" } }
        }
      },
      "Offering": { "allOf": [{ "type": "object", "properties": { "id": { "type": "string" } } }, { "$ref": "#/components/schemas/OfferingCreate" }] },
      "ExperimentCreate": {
        "type": "object",
        "required": ["name", "control_offering_id", "treatment_offering_id", "traffic_pct"],
        "properties": {
          "name": { "type": "string" },
          "status": { "type": "string", "enum": ["draft", "running", "stopped"] },
          "control_offering_id": { "type": "string" },
          "treatment_offering_id": { "type": "string" },
          "traffic_pct": { "type": "integer", "minimum": 0, "maximum": 100 }
        }
      },
      "Experiment": { "allOf": [{ "type": "object", "properties": { "id": { "type": "string" }, "created_at": { "type": "string", "format": "date-time" } } }, { "$ref": "#/components/schemas/ExperimentCreate" }] },
      "ExperimentResults": {
        "type": "object",
        "properties": {
          "control": { "$ref": "#/components/schemas/VariantResult" },
          "treatment": { "$ref": "#/components/schemas/VariantResult" }
        }
      },
...
      "App": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "public_api_key": { "type": "string" },
          "bundle_id": { "type": ["string", "null"] },
          "package_name": { "type": ["string", "null"] },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
...
    }
  }
}